Split into files and move svg to external one
